requirements.txt
aniso8601==8.1.1
appdirs==1.4.4
argon2-cffi==20.1.0
asgiref @ file:///tmp/build/80754af9/asgiref_1617740417847/work
autopep8==1.5.7
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
distlib==0.3.2
Django @ file:///tmp/build/80754af9/django_1617740316746/work
django-bootstrap3==15.0.0
django-bootstrap4==3.0.1
django-braces==1.14.0
django-cors-headers==3.7.0
django-debug-toolbar==3.2.1
django-environ==0.4.5
django-graphql-jwt==0.3.2
django-rest-auth==0.9.5
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.0
filelock==3.0.12
graphene==3.0b7
graphene-django==3.0.0b7
graphql-core==3.1.5
graphql-relay==3.1.0
idna==2.10
lxml==4.6.3
Pillow==8.2.0
promise==2.3
psycopg2 @ file:///opt/concourse/worker/volumes/live/7541fe58-2a45-4073-5d69-313e733fc729/volume/psycopg2_1612298592535/work
pycodestyle==2.7.0
pycparser==2.20
PyJWT==2.1.0
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz @ file:///tmp/build/80754af9/pytz_1612215392582/work
requests==2.25.1
Rx==1.6.1
simplejson==3.17.2
singledispatch==3.6.2
six==1.15.0
soupsieve==2.2.1
sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
text-unidecode==1.3
toml==0.10.2
urllib3==1.26.4
virtualenv==20.4.7
virtualenv-clone==0.5.4
